requirements_doppleganger.txt
# Core Flask Application & Web Framework
Flask>=2.0.0
Flask-SQLAlchemy>=2.5.0
Flask-Migrate>=3.0.0
Flask-WTF>=1.0.0
Flask-Login>=0.5.0
Flask-Bcrypt>=1.0.0
Flask-Mail>=0.9.0
python-dotenv>=0.19.0
gunicorn>=20.0.0
Flask-Cors>=3.0.10

# Face Matching & Image Processing
faiss-cpu>=1.7.0
face_recognition>=1.3.0
dlib>=19.22.0
Pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0

# Other Utilities
concurrent-log-handler>=0.9.20
redis>=4.0.0  # Added for Redis cache/broker support
Flask-Caching>=2.0.0  # Added for caching support
Flask-Limiter>=3.0.0  # Added for rate limiting
Flask-Session>=0.4.0  # Added for server-side sessions
Flask-Talisman>=1.0.0  # For security headers
Flask-Moment>=0.11.0  # For formatting dates and times
PyJWT>=2.0.0  # For JSON Web Token handling
requests>=2.25.0  # For making HTTP requests
tqdm>=4.60.0  # For progress bars
